requirements.txt
requests
edge-tts
moviepy
Pillow
python-dotenv
